requirements.txt
antlr4-python3-runtime==4.8
args==0.1.0
boto3==1.18.7
botocore==1.21.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
Cython==0.29.24
-e git+https://github.com/pytorch/fairseq.git@67ff6baa42c1208d0da85f5af2f01689034d1dfd#egg=fairseq
hydra-core==1.0.7
idna==3.2
importlib-metadata==4.6.1
importlib-resources==5.2.0
jmespath==0.10.0
joblib==1.0.1
numpy==1.21.1
omegaconf==2.0.6
pandas==1.3.1
portalocker==2.0.0
pycparser==2.20
python-dateutil==2.8.2
pytorch-transformers==1.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
s3transfer==0.5.0
sacrebleu==1.5.1
sacremoses==0.0.45
sentencepiece==0.1.96
six==1.16.0
torch==1.9.0
torchtext==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
